feat/Сверстал компонент promo
