+++ src/components/header/Header.js
@@ -14,36 +14,34 @@ import MobuleMenu from './mobule-menu/MobuleMenu';
 // Пользовательские хуки. HOC не подойдет
 import useMobuleMenu from '../../utils/custom-hooks/useMobuleMenu';
 
+// constants
+import {
+  textList,
+  navLink,
+} from '../../utils/constants';
+
 function Header() {
   // Принимает минимальную ширину экрана для мобильной версии
   const isMobuleMenu = useMobuleMenu(769);
 
-  const textList = [
-    'Фильмы',
-    'Сохранённые фильмы',
-  ];
-
-  const navLink = [
-    '/main',
-    '/main1',
-  ];
-
   return (
-    <header className='header page__margin-auto'>
-      <Link to='/main' className='header__link-logo' >
-        <img src={ logo } alt='Логотип' className='header__logo' />
-      </Link>
-      {/* <AuthLink /> */}
-      {
-        isMobuleMenu ?
-          <MobuleMenu />
-        :
-          <>
-            <Navigation textList={ textList } navLink={ navLink } />
-            <ProfileLink />
-          </>
-      }
-    </header>
+    <div className='background'>
+      <header className='header page__margin-auto'>
+        <Link to='/main' className='header__link-logo' >
+          <img src={ logo } alt='Логотип' className='header__logo' />
+        </Link>
+        {/* <AuthLink /> */}
+        {
+          isMobuleMenu ?
+            <MobuleMenu />
+          :
+            <>
+              <Navigation textList={ textList } navLink={ navLink } />
+              <ProfileLink />
+            </>
+        }
+      </header>
+    </div>
   );
 }
