+++ src/components/header/mobule-menu/MobuleMenu.js
@@ -1,9 +1,21 @@
-import React, { useState } from 'react';
+import React, { /* useEffect, */ useState } from 'react';
 import './mobuleMenu.css';
 
 // Компоненты
 import ProfileLink from '../profile-link/ProfileLink';
 import Navigation from '../navigation/Navigation';
+import HeaderToggleMenu from './header-toggle-menu/HeaderToggleMenu';
+
+// constants
+import {
+  textListMenuMobule,
+  navLinkMenuMobule,
+  modifierNavigationClass,
+  modifierProfileLink,
+} from '../../../utils/constants';
+
+// utile
+// import closeMenuOnEsc from '../../../utils/closeMenuOnEsc';
 
 function MobuleMenu() {
   const [ activeMenu, setActiveMenu ] = useState(false);
@@ -18,13 +30,24 @@ function MobuleMenu() {
   'mobule-menu__content mobule-menu__content_active' :
   'mobule-menu__content';
 
-  // Закрытие меню по нажатию на esc
-  function closeMenuOnEsc(e) {
-    if (activeMenu && e.key === 'Escape') {
-      setActiveMenu(false)
-    }
-  }
-  document.addEventListener('keydown', closeMenuOnEsc);
+  // const closeMenuOnEsc = e => {
+  //   console.log(1)
+  //   if (activeMenu && e.key === 'Escape') {
+  //     setActiveMenu(false)
+  //   }
+  // }
+
+  // useEffect(() => {
+  //   console.log(activeMenu)
+  //   return () => {
+  //     if (activeMenu === false) {
+  //       document.removeEventListener('keydown', closeMenuOnEsc);
+  //       return;
+  //     }
+  //     console.log(10)
+  //     document.addEventListener('keydown', closeMenuOnEsc);
+  //   }
+  // }, [ activeMenu ]);
 
   // Закрытие меню по нажатию на контейнер
   function closeMenuOnContainer(e) {
@@ -33,55 +56,17 @@ function MobuleMenu() {
     }
   }
 
-  const textList = [
-    'Главная',
-    'Фильмы',
-    'Сохранённые фильмы',
-  ];
-
-  const navLink = [
-    '/main1',
-    '/main',
-    '/main3',
-  ];
-
-  // Необязательный обьект для навигации. Устанавливает модификаторы
-  const modifierNavigationClass = {
-    navigationClass: 'header__navigation_position_mobule-menu',
-    navigationListClass: 'header__navigation-list_position_mobule-menu',
-    navigationItemClass: 'header__navigation-item_position_mobule-menu',
-    navigationLinkActive: 'header__navigation-link_mobule-menu_active',
-  };
-
-  const modifierProfileLink = {
-    profileLink: 'header__profile-link_position_mobule-menu',
-  };
-
-  const toggleMenu = () => {
-    setActiveMenu(!activeMenu);
-  };
-
-  const activeMenuSetClass = (activeClass) => {
-    if ( activeMenu ) {
-      return `header__toggle-menu-item ${activeClass}`;
-    }
-    return 'header__toggle-menu-item';
-  };
-
   return (
     <>
-      {/* menu-toggle */}
-      <button type='button' className='header__toggle-menu' onClick={ toggleMenu }>
-          <ul className='header__toggle-menu-list'>
-            <li className={ activeMenuSetClass('header__toggle-menu-item_rotate_right') }></li>
-            <li className={ activeMenuSetClass('header__toggle-menu-item_active') }></li>
-            <li className={ activeMenuSetClass('header__toggle-menu-item_rotate_left') }></li>
-          </ul>
-        </button>
+      <HeaderToggleMenu setActiveMenu={ setActiveMenu } activeMenu={ activeMenu } />
       <div className={ mobuleMenuClass }>
         <div className={ mobuleMenuContainerClass } onClick={ closeMenuOnContainer }>
           <div className={ mobuleMenuContentClass }>
-            <Navigation textList={ textList } modifierNavigationClass={ modifierNavigationClass } navLink={ navLink } />
+            <Navigation
+              textList={ textListMenuMobule }
+              modifierNavigationClass={ modifierNavigationClass }
+              navLink={ navLinkMenuMobule }
+            />
             <ProfileLink modifierProfileLink={ modifierProfileLink } />
           </div>
         </div>
