+++ src/components/header/mobule-menu/mobuleMenu.css
@@ -1,59 +1,3 @@
-.header__toggle-menu {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  max-width: 40px;
-  width: 100%;
-  min-height: 43px;
-
-  padding: 0;
-  cursor: pointer;
-  background-color: inherit;
-  border: none;
-
-  z-index: 1;
-}
-
-.header__toggle-menu-list {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-
-  margin: 0;
-  padding: 0;
-  list-style: none;
-}
-
-.header__toggle-menu-item {
-  background-image: url(../../../images/svg/header-toggle-menu-icon.svg);
-  background-repeat: no-repeat;
-  background-position: center;
-  background-size: cover;
-  width: 22px;
-  height: 2.5px;
-  margin-bottom: 6.5px;
-
-  transition: transform .3s ease, opacity .3s ease;
-}
-
-.header__toggle-menu-item:last-of-type {
-  margin-bottom: 0px;
-}
-
-.header__toggle-menu-item_rotate_right {
-  transform: rotate(45deg) translateX(6px) translateY(6px);
-}
-
-.header__toggle-menu-item_rotate_left {
-  transform: rotate(-45deg) translateX(6px) translateY(-6px);
-}
-
-.header__toggle-menu-item_active {
-  visibility: hidden;
-  opacity: 0;
-}
-
 .mobule-menu {
   position: fixed;
   top: 0;
