+++ src/components/main/Main.js
@@ -4,14 +4,15 @@ import './main.css';
 // Компоненты
 import Header from '../header/Header';
 import Footer from '../footer/Footer';
+import Promo from './promo/Promo';
 
 function Main() {
   return (
     <>
       <Header />
-      <h1>
-        Main
-      </h1>
+      <main className='about-project'>
+        <Promo />
+      </main>
       <Footer />
     </>
   );
