--- src/components/main/nav-tab/NavTab.js
@@ -1,12 +1,0 @@
-import React from 'react';
-import './navTab.css';
-
-function NavTab() {
-  return (
-    <h1>
-      NavTab
-    </h1>
-  );
-}
-
-export default NavTab;
