+++ src/components/main/promo/Promo.js
@@ -1,11 +1,22 @@
 import React from 'react';
 import './promo.css';
 
+import NavTab from './nav-tab/NavTab';
+
 function Promo() {
   return (
-    <h1>
-      Promo
-    </h1>
+    <div className='background'>
+      <section className='promo page__margin-auto'>
+        <h1 className='promo__title'>
+          Учебный проект студента факультета
+          <br/>Веб-разработки.
+        </h1>
+        <p className='promo__desc'>
+          Листайте ниже, чтобы узнать больше про этот проект и его создателя.
+        </p>
+        <NavTab />
+      </section>
+    </div>
   );
 }
