+++ src/components/main/promo/nav-tab/navTab.css
@@ -1,0 +1,20 @@
+.promo__navigation {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  max-width: 130px;
+  width: 100%;
+  min-height: 36px;
+  box-sizing: border-box;
+
+  text-decoration: none;
+  background-color: inherit;
+  border: 1px solid black;
+  border-radius: 3px;
+
+  font-family: Inter;
+  font-weight: 500;
+  font-size: 12px;
+  line-height: 16px;
+  color: black;
+}
