+++ src/components/main/promo/promo.css
@@ -1,0 +1,42 @@
+.background {
+  width: 100%;
+  background-color: #F3C1F8;
+}
+
+.promo {
+  max-width: 1140px;
+  padding: 0 10px;
+  box-sizing: border-box;
+
+  background-image: url(../../../images/promo-text-color.svg);
+  background-repeat: no-repeat;
+  background-size: 310px 304px;
+  background-position: right 75px;
+
+  padding-bottom: 75px;
+}
+
+.promo__title {
+  max-width: 640px;
+  margin: 0;
+  padding-top: 75px;
+  margin-bottom: 36px;
+
+  font-family: Inter;
+  font-size: 50px;
+  font-weight: 400;
+  line-height: 58px;
+  color: black;
+}
+
+.promo__desc {
+  max-width: 310px;
+  margin: 0;
+  margin-bottom: 146px;
+
+  font-family: Inter;
+  font-weight: 400;
+  font-size: 13px;
+  line-height: 16px;
+  color: rgba(0, 0, 0, .5);
+}
